chore: remove unused value_functions
